Rename graphqlapi folder into proxy
